consolidate APIs into apis.cpp
