+++ src/gdsync.cpp
@@ -536,18 +536,7 @@ out:
   INLCPY 128B
 */
 
-static inline uint32_t gds_qword_lo(uint64_t v) {
-        return (uint32_t)(v);
-}
-static inline uint32_t gds_qword_hi(uint64_t v) {
-        return (uint32_t)(v >> 32);
-}
-
-enum {
-        GDS_POST_OPS_DISCARD_WAIT_FLUSH = 1<<0
-};
-
-static int gds_post_ops(size_t n_ops, struct peer_op_wr *op, CUstreamBatchMemOpParams *params, int &idx, int post_flags = 0)
+int gds_post_ops(size_t n_ops, struct peer_op_wr *op, CUstreamBatchMemOpParams *params, int &idx, int post_flags)
 {
         int retcode = 0;
         size_t n = 0;
@@ -1521,171 +1510,6 @@ int gds_query_param(gds_param_t param, int *value)
                 ret = EINVAL;
                 break;
         };
-        return ret;
-}
-
-//-----------------------------------------------------------------------------
-
-static bool no_network_descs_after_entry(size_t n_descs, gds_descriptor_t *descs, size_t idx)
-{
-        bool ret = true;
-        size_t i;
-        for(i = idx+1; i < n_descs; ++i) {
-                gds_descriptor_t *desc = descs + i;
-                switch(desc->tag) {
-                case GDS_TAG_SEND:
-                case GDS_TAG_WAIT:
-                        ret = false;
-                        goto out;
-                case GDS_TAG_WAIT_VALUE32:
-                case GDS_TAG_WRITE_VALUE32:
-                        break;
-                default:
-                        gds_err("invalid tag\n");
-                        ret = EINVAL;
-                        goto out;
-                }
-        }
-out:
-        return ret;
-}
-
-static int get_wait_info(size_t n_descs, gds_descriptor_t *descs, size_t &n_waits, size_t &last_wait)
-{
-        int ret = 0;
-        size_t i;
-        for(i = 0; i < n_descs; ++i) {
-                gds_descriptor_t *desc = descs + i;
-                switch(desc->tag) {
-                case GDS_TAG_WAIT:
-                        ++n_waits;
-                        last_wait = i;
-                        break;
-                case GDS_TAG_SEND:
-                case GDS_TAG_WAIT_VALUE32:
-                case GDS_TAG_WRITE_VALUE32:
-                        break;
-                default:
-                        gds_err("invalid tag\n");
-                        ret = EINVAL;
-                }
-        }
-        return ret;
-}
-
-static size_t calc_n_mem_ops(size_t n_descs, gds_descriptor_t *descs)
-{
-        size_t n_mem_ops = 0;
-        size_t i;
-        for(i = 0; i < n_descs; ++i) {
-                gds_descriptor_t *desc = descs + i;
-                switch(desc->tag) {
-                case GDS_TAG_SEND:
-                        n_mem_ops += desc->send->commit.entries + 2; // extra space, ugly
-                        break;
-                case GDS_TAG_WAIT:
-                        n_mem_ops += desc->wait->peek.entries + 2; // ditto
-                        break;
-                case GDS_TAG_WAIT_VALUE32:
-                case GDS_TAG_WRITE_VALUE32:
-                        n_mem_ops += 2; // ditto
-                        break;
-                default:
-                        gds_err("invalid tag\n");
-                }
-        }
-        return n_mem_ops;
-}
-
-int gds_stream_post_descriptors(CUstream stream, size_t n_descs, gds_descriptor_t *descs)
-{
-        size_t i;
-        int idx = 0;
-        int ret = 0;
-        int retcode = 0;
-        size_t n_mem_ops = 0;
-        size_t n_waits = 0;
-        size_t last_wait = 0;
-        bool move_flush = false;
-
-        n_mem_ops = calc_n_mem_ops(n_descs, descs);
-        get_wait_info(n_descs, descs, n_waits, last_wait);
-
-        gds_dbg("n_descs=%zu n_waits=%zu n_mem_ops=%zu\n", n_descs, n_waits, n_mem_ops);
-
-        // move flush to last wait in the whole batch
-        if (n_waits && no_network_descs_after_entry(n_descs, descs, last_wait)) {
-                gds_dbg("optimizing FLUSH to last wait i=%zu\n", last_wait);
-                move_flush = true;
-        }
-        // alternatively, remove flush for wait is next op is a wait too
-
-        CUstreamBatchMemOpParams params[n_mem_ops];
-
-        for(i = 0; i < n_descs; ++i) {
-                gds_descriptor_t *desc = descs + i;
-                switch(desc->tag) {
-                case GDS_TAG_SEND: {
-                        gds_send_request_t *sreq = desc->send;
-                        retcode = gds_post_ops(sreq->commit.entries, sreq->commit.storage, params, idx);
-                        if (retcode) {
-                                gds_err("error %d in gds_post_ops\n", retcode);
-                                ret = retcode;
-                                goto out;
-                        }
-                        // TODO: fix late checking
-                        //assert(idx <= n_mem_ops);
-                        if (idx >= n_mem_ops) {
-                                gds_err("idx=%d is past allocation (%zu)\n", idx, n_mem_ops);
-                                assert(!"corrupted heap");
-                        }
-                        break;
-                }
-                case GDS_TAG_WAIT: {
-                        gds_wait_request_t *wreq = desc->wait;
-                        int flags = 0;
-                        if (move_flush && i != last_wait)
-                                flags = GDS_POST_OPS_DISCARD_WAIT_FLUSH;
-                        retcode = gds_post_ops(wreq->peek.entries, wreq->peek.storage, params, idx, flags);
-                        if (retcode) {
-                                gds_err("error %d in gds_post_ops\n", retcode);
-                                ret = retcode;
-                                goto out;
-                        }
-                        // TODO: fix late checking
-                        assert(idx <= n_mem_ops);
-                        break;
-                }
-                case GDS_TAG_WAIT_VALUE32:
-                        retcode = gds_fill_poll(params+idx, desc->wait32.ptr, desc->wait32.value, desc->wait32.cond_flags, desc->wait32.flags);
-                        if (retcode) {
-                                gds_err("error %d in gds_fill_poll\n", retcode);
-                                ret = retcode;
-                                goto out;
-                        }
-                        ++idx;
-                        break;
-                case GDS_TAG_WRITE_VALUE32:
-                        retcode = gds_fill_poke(params+idx, desc->write32.ptr, desc->write32.value, desc->write32.flags);
-                        if (retcode) {
-                                gds_err("error %d in gds_fill_poke\n", retcode);
-                                ret = retcode;
-                                goto out;
-                        }
-                        ++idx;
-                        break;
-                default:
-                        assert(0);
-                        break;
-                }
-        }
-        retcode = gds_stream_batch_ops(stream, idx, params, 0);
-        if (retcode) {
-                gds_err("error in batch_ops\n");
-                goto out;
-        }
-
-out:
         return ret;
 }
